Reorg platform/web src directory
