+++ src/platform/web/modifiers/index.ts
@@ -1,0 +1,5 @@
+/**
+ * SessionDescriptionHandlerModifer functions for web browsers.
+ * @packageDocumentation
+ */
+export * from "./modifiers";
