+++ src/platform/web/session-description-handler/index.ts
@@ -1,0 +1,7 @@
+/**
+ * A SessionDescriptionHandler for web browsers.
+ * @packageDocumentation
+ */
+export * from "./peer-connection-options";
+export * from "./session-description-handler";
+export * from "./session-description-handler-options";
