+++ src/platform/web/session-description-handler/peer-connection-options.ts
@@ -1,0 +1,8 @@
+/**
+ * Options for PeerConnection.
+ * @public
+ */
+export interface PeerConnectionOptions {
+  iceCheckingTimeout?: number;
+  rtcConfiguration?: RTCConfiguration;
+}
