+++ src/platform/web/session-description-handler/session-description-handler-options.ts
@@ -1,0 +1,14 @@
+import { SessionDescriptionHandlerOptions as SessionDescriptionHandlerOptionsDefinition } from "../../../api";
+import { PeerConnectionOptions } from "./peer-connection-options";
+
+/**
+ * Options for {@link SessionDescriptionHandler}.
+ * @public
+ */
+export interface SessionDescriptionHandlerOptions extends SessionDescriptionHandlerOptionsDefinition {
+  peerConnectionOptions?: PeerConnectionOptions;
+  alwaysAcquireMediaFirst?: boolean;
+  disableAudioFallback?: boolean;
+  RTCOfferOptions?: any;
+  constraints?: MediaStreamConstraints;
+}
