+++ src/platform/web/transport/index.ts
@@ -1,0 +1,6 @@
+/**
+ * A Transport implementation for web browsers.
+ * @packageDocumentation
+ */
+export * from "./transport";
+export * from "./transport-options";
